Rename to match botocore
I feel like maybe it should be "parsing" and "serialization" but we'll keep the botocore names for now...
