+++ moto/core/responses.py
@@ -34,7 +34,7 @@ from werkzeug.exceptions import HTTPException
 from moto import settings
 from moto.core.common_types import TYPE_IF_NONE, TYPE_RESPONSE
 from moto.core.exceptions import DryRunClientError, ServiceException
-from moto.core.parse import PROTOCOL_PARSERS, XFormedDict
+from moto.core.parsers import PROTOCOL_PARSERS, XFormedDict
 from moto.core.request import normalize_request
 from moto.core.serialize import SERIALIZERS, ResponseSerializer, XFormedAttributePicker
 from moto.core.utils import (
